initialise_gauss_profilers moved to profiler_gauss
